-created an alternate SagaOf (MagicSagaOf)
-created an alternate SagaNotificationHandler (MagicSagaNotificationHandler)
-created an alternate RegisterSaga helper (MagicRegisterSaga)
-created a HetrogenousContainer for storing multiple idmappings
-removed content mapping code, as this now happens at the handler level

created tests for the variations in the 'magic' implementations to verify everything still functions, just the how has changed

the term magic is prefixed so i can keep the existing implementations whilst testing this new implementation. it is not really magic in any way

all this amounts to using the handles method for performing content mapping, as i believe this allows for more flexible and idiotmatic handling of events in the saga handlers. content can still naively be mapped in these methods, but it seems more correct to have the handles methods 'handle' the events, especially when the content mapping is more complex than a simple member = value assignment. the TryComplete concept still exists, however this is now executed from the sagaHandler, when the saga is no longer in the pending state. this is the same behaviour as the incumbent solution, however it also seems more straight forward, as the handles methods dont need to all blindly call TryComplete. instead, the TryComplete is called once, and is only called under the same condtion that is is called under when using the Handles method to call TryComplete. Effectively this means the behaviour hasn't (or shouldnt have changed), only the implemenation, and *where* the event handling is performed (in the handles method, rather than the content mapping lambdas)

i would also like to re-arrange how the id mappings are registered, as there is currently a potential bug where the abstract member is called from the ctor of the base. if the deriving saga is not sealed, and further derivations are created, this could easily create a bug due to the order of ctor calls according to the CLR spec. Ill leave this for another day though.